device: Make device.Manager embed device.Device
For management systems that we want to get information from
both the system itself and the devices it manages, they need to
implement both a Device and Manager to achieve this, which is a bit
confusing. This commit makes Manager embed Device so that only
one implementation is needed.

Tested by pointing the Collector to real devices.

Change-Id: If976f8ffeb5083c83599e24f6def299437b5d8cc