device: Add Pattern field to device.Option
I've added a Pattern field to the device.Option, which allows (but
doesn't require) device implementations to specify a regular expression
that valid option values must match. These patterns are checked in
sanitizedConfig, transparently to device implementations, so all device
implementations must check for is combinations of options that are
invalid in an application-specific way.

Change-Id: I0ae61a016c12354ba8cf7b70e673e14877cab9e6