provider/snmp: Handle SNMP packet errors
The SNMP packet contains an error field that we're currently ignoring.
I've updated the packet handling logic to look for an error and
translate the code into something human-readable, and return an error.

I tested this against a real SNMP agent, giving v3 options that resulted
in a noAccess error.

Change-Id: I734e35246648162e41bad00c43e2352c5005de72