provider/snmp: Add /system/state/boot-time
I've added boot time to updateSystemState. We first try for
snmpEngineTime, which is an int32 representing the number of seconds
since boot, and then sysUpTimeInstance, which is an int32 representing
hundredths of a second since boot.

I added a couple of test cases to snmp_test.go and I tested manually
against an EOS dut + CVP dut to verify that uptime shows up in Geiger.

Change-Id: Iab103bd2d973aebf85554148f629688fee7af9e0