misc: Replace glog with logrus and support log multiplexing
Added a logging infrastructure to redirect all logs into a directory,
one file per device. device.LogDevice() and provider.LogProvider() are
the wrappers that should be used if one wants the logs to be
multiplexed.

Tested by running the Collector with multiple devices with -logDir and
seeing logs are written in that many files.

Change-Id: Ia70c7f99ceb0cf21c6e64c6569e771f33817ac33