version: Add version.CollectorVersion
version.CollectorVersion allows us to distinguish the version of a built
Collector from the version of the cloudvision-go repo used to build that
Collector. This is handy for building a Collector that contains code
that's not in the cloudvision-go repo. Whenever we see a version.Version
that's different than version.CollectorVersion, we should have -version
print out both. Collectors built from cloudvision-go should have
version.Version == version.CollectorVersion.

Change-Id: I3cc72bbcff390edd113d7f03b3ae75ba6c1e8507