all: Support grpc for CRUD operations on device
Add a grpc service to do CRUD on devices as an alternative way to modify
the inventory on the fly than the dynamic configration.

Tried running it in all three modes. There's still room for optimization
like how/when do we do backup/restore, but send out an early review now.

Change-Id: I3b6b4530f855d28b21ea13ffbac61ade4cb54eab