provider/snmp: Update SNMP provider to use qualified form for YANG types
Due to some changes being made in gopenconfig, some YANG identity names
need to be expressed in qualified form. We're seeing Collector failures
because of this. I've verified with the adapter tests that this fix does
what we want--once this code is merged I'll pull these changes into
arista and then update the integration test that wasn't catching this.

Change-Id: I016b6403c9edbf58bbf7145c9ce2ff3ebd008b02