provider/snmp: Add SNMP - OpenConfig translator
I've added a framework for doing SNMP - OpenConfig translations. It's
basically a work in progress at this point--I need to think through
the way gNMNI indexes interact with SNMP indexes a bit more. In
particular, there needs to be a way for a mapping for a path
a/b[k=v]/c to produce the prefix path a/b[k=v] efficiently. In some
cases these index mappings will be complex and have fallback logic, and
we need to figure out how to handle those cases.

Change-Id: Idcb3f9b56ee907f08b5901c53e084fb64fae5743