libmain: Remove backupFile flag from Collector
Removed everything that directly/indirectly used the backupFile flag
because it is not good to have code that is not going to be supported.

backupFile was used as a parameter in many functions, so those
parameters were removed because it was not being used. If backupFile was
returned in a function, it was replaced with nil.

In inventory.go, backup() and restore() functions were removed because
it will be done in other parts of the system not with the backupFile
flag.

The os package in the inventory.go imports was removed because it was
not used.

Change-Id: I6feb369564fc032ae8c382562793a38b8e09152d