provider/snmp: Remove square brackets in gNMI strings
We've been letting '[' and ']' appear in gNMI string values, which
confuses the gNMI path parser. I've changed the string sanitizer to
replace square brackets with parentheses. We might be able to escape the
square brackets (this satisfies the path parser) but the escaped
brackets then cause trouble in the unmarshaling. Using parentheses is
easier and I don't think it will break anything.

Change-Id: Ibe37b0866b2866668d2a3fd8ac7e3a3b1aeff1fd