smi/pdu: Handle OID indexes that span multiple OID components
As long as it can identify the base object type, smi.Store.GetObject
shouldn't make any fuss if it looks like an OID has too many indexes.

I've also added a method to the pdu.Store, IndexValueByName, that
returns the value of a specified index in a PDU's OID. It doesn't fully
understand how to find an index--it assumes one index per OID
component--but we can teach it that when we need to.

Change-Id: I1842a014d9979ca43681c1b6391f1ba0b4dbe608