smi: Handle orphan objects
Sometimes an object is declared in a different module than its parent.
For example, sysUpTimeObject, from DISMAN-EVENT-MIB, is a child of
sysUpTime, from SNMPv2-MIB. In such cases, since we don't necessarily
know the parent object at the time the child object is first discovered,
in the resolution phase we need to link the two. I've added a test case
to check that we're adding parents for these objects.

Change-Id: I9d126e156af3afb3ee7059a32b83e110121ed18a