snmpoc: Prepare translator for use by snmp provider
The purpose of this commit is to get the translator ready for the snmp
provider to use it. I've copied all the snmp provider tests into
translator_test, and changed the translator interface a bit.

The most notable change is that the translator now takes a gosnmp object
and does the SNMP operations itself rather than expecting the caller to
prepare for the translation. The translator, which contains the mappings,
already has to know about dependencies between SNMP data and OpenConfig
paths, and between OpenConfig paths, so it's a better place to determine
the ordering of polls and translations. Ideally this information can be
built into the mappings themselves, but for now it's hard-coded into the
translator code.

Change-Id: I409099b97ec746547b95348c82060c9cd54decab