provider/snmp: Handle unsigned SNMP subtype values
The value of lldpLocChassisIdSubtype has type INTEGER, but we've seen
unsigned values in the wild. This results in a panic in code that
assumes ints, so I've updated all the code that gets a subtype value and
converts it to a string to sanitize the subtype value to an int. I've
added a test case for both the lldpChassisID OpenConfig mapping and for
the DeviceID code.

Change-Id: I490ad7d612eaeedbf9de35e0453d705a5dedffa8