provider/snmp: Add dummy Go package to provider/snmp/smi/mibs
In order to get the MIBs past dep, I've created a dummy Go package in
the mibs directory and had the smi package import it. I also had to add
logic to the SMI parser to have it skip files with extensions that
aren't ".mib" so that it doesn't try to parse this dummy Go file.

Change-Id: Id916a2cea403cd8acb79d74afee5d26b7e72def0