provider/snmp: LLDP translation fixes
This commit contains fixes for two bugs discovered on customer setups:

(1) Handle non-scalar OIDs ending in ".0", i.e., allow indexes to be
'0'. We hadn't seen this before but it's valid, so we shouldn't assume
anything ending in '.0' is a scalar.

(2) Don't give up in LLDP translation when something goes wrong on one
interface. During our lldpRemTable walk e're seeing a couple of local
interface names that don't have corresponding entries in ifTable (and in
fact that conflict with corresponding interfaces indexes). This
shouldn't stop us from issuing updates on the other interfaces. So now
we log and continue. This required adding logging to the mappers, which
is what most of the code in this review is doing.

Change-Id: I663580502b3e132b1d66ed8d8bf1144ef18bb3fc