device: Refactor code to abstract out communication with CV
As the first step in supporting communication of the collector
with CV via the YAIL layer, this commit abstracts out all gNMI
adapter specific details and hides them behind a CVClient
interface. Then the cvclient/v1 package implements this interface
to allow communication with the gNMI adapter. As the next step,
we will implement a v2 implementation of the CVClient interface
for communicating via the YAIL based service in CV.

fixes=BUG493304

Change-Id: I3ccd11da8941b8097fc298599a7ce5f3ea2a1e40