libmain: Add command line flags for authenticating to gNMI service on CV
This change adds support for command line flag to specify access token
and optionally server CA cert file for secure, authenticated
communication with the gNMI service on CVP/CVaas. This is needed when
the collector is not running on CV itself.

Also fixed a bug in auth.go where the format of the authorization header
carrying the access token was incorrect.

Change-Id: I051bbb83b5342a7452036ace60c0bcf9d8d0cda9