provider: Add GRPCProvider
I've added another provider type, GRPCProvider, which is a Provider
that is given a raw gRPC connection. I've also updated
device/inventory.go to check whether a device's provider is a
GRPCProvider, and if so to call InitGRPC on it.

The Collector libmain also had to be updated to take an optional
-grpcServerAddr option and connect to that address if provided.

Change-Id: I4f81bd8474387573952beaa11f2c85ab95c658ad