provider/snmp: Use safe typecast, handle unexpected PDU type
It's possible to get back a PDU with a different OID than the one
requested. We need to be able to handle this by detecting the bad type,
logging, and carrying on.

Separately, the reason we've been seeing this case is on a v3 auth
error. We should handle these errors in snmpNetworkInit and
short-circuit device creation, since they're unrecoverable. This change
adds such a check.

Change-Id: I2329ca70e566bf2f0eb97ee3f9000c22217475ac