provider/snmp/snmpoc: Map SNMP entPhysicalModelName to OC part-no
As of https://gerrit.corp.arista.io/c/arista/+/199911 we're versioning
the OpenConfig path /components/component/state/part-no to
/Devices/<device>/versioned-data/Device/modelName, but we haven't
updated the SNMP mappings to use part-no yet (we're still using
hardware-revision mistakenly). This change updates the provider to
translate to part-no, which will allow us to show the model name in the
inventory correctly again.

Change-Id: I4e07652f9dfba25130ac860a6b92795ea8a8f29d