device/openconfig: Add timeout to openConfig device Alive call
Without access to the device context and no timeout, we may fail to
decommission the device if the connection is halted for a large
amount of time. Setting the timeout to 10s until we refactor
to take in the request context that will allow immediate cancelation.

Change-Id: I645e5199d0a7ca5e249de3267b6555ed8e330b93
(cherry picked from commit c7f578fd3d28b01f6fee7ec70d00e39d57a58bca)