sensor: Add message when the device is back alive
When error happens that cause the device to be offline a message is
logged. However, when the device comes back online no new message
is written, and the message log will look like the device is
still having issues. So we should log a message when the device
returns to a stable state.

Also including mock library and generating some mock interfaces
to simplify testing logic to move away from custom testing clients.

Change-Id: I004336f2f616afb361f6edb170bddfd75207ac19