device: enforce cluster clock sync between CVaaS and Sensor
CVaaS cluster clock and sensor clock should be in sync, so that
notification sent to CVaaS from sensor will have correct timestamp.
If these clocks are not in sync then CVaaS might receive future or
past dated notifications, to avoid this we added new interface and
method which will post status of clock on channel and sensor
will stop/start sensor and datasource streaming based on that.

Closes: BUG829844

Change-Id: I4965cd89315fb6c44fddaf21b401eea6dadd757a