api/examples: Add script to run arbitrary actions and args in Go
Added a Go conversion of the run_action.py
script from cloudvision-python to serve
as an example conversion from python
to GoLang, as well as to provide a
tool written entirely in Go capable of
packaging and running arbitrary actions
and args as change controls.
This functionality should be useful for
future features in the arista go repo,
such as the action triggers by cv events
feature currently in progress.

Change-Id: Ic414d4429b85fda423897094fa57f6b62bb08cdf